feat: Updated receipts table

+++ .gen/public/model/receipts.go
@@ -12,7 +12,7 @@ import (
 )
 
 type Receipts struct {
-	ID          int32 `sql:"primary_key"`
+	ID          int64 `sql:"primary_key"`
 	UserID      int64
 	Description string
 	URLSlug     string
